new stats

+++ testseries/mean.py
@@ -8,12 +8,12 @@ if len(sys.argv) < 2:
     sys.exit()
 
 j = 0
-i = 0
+i = 0.0
 with open(sys.argv[1]) as f:
     binWriter = csv.writer(open(sys.argv[1] + ".csv", 'ab'))
     for line in f:
         binWriter.writerow(line)
-        i = i + int(line)
+        i = i + float(line)
         j += 1
 
 print i/j
